moved caesar cipher calling methods to its own file & created basic test for rspec
