+++ caesar_cipher.rb
@@ -34,10 +34,3 @@ def is_normal_character?(letter)
   is_small_letter || is_capital_letter
 end
 
-puts caesar_cipher("xyz!", 3)
-puts caesar_cipher("abc!", -3)
-puts caesar_cipher("XYZ!", 3)
-puts caesar_cipher("ABC!", -3)
-puts caesar_cipher("What a string!", 5)
-puts caesar_cipher("What a string!", 31)
-#Bmfy f xywnsl!
