+++ caesar_cipher_main.rb
@@ -1,0 +1,8 @@
+require_relative 'caesar_cipher'
+puts caesar_cipher("xyz!", 3)
+puts caesar_cipher("abc!", -3)
+puts caesar_cipher("XYZ!", 3)
+puts caesar_cipher("ABC!", -3)
+puts caesar_cipher("What a string!", 5)
+puts caesar_cipher("What a string!", 31)
+#Bmfy f xywnsl!
