+++ spec/caesar_cipher_spec.rb
@@ -1,0 +1,7 @@
+require './caesar_cipher.rb'
+
+describe '#caesar_cipher' do
+  it "Is ok" do
+    expect(true).to eq (true)
+  end
+end
